removing LICENSE, adding config file check

+++ kanjied.rb
@@ -1,25 +1,4 @@
-##!/bin/env ruby
-#Copyright (c) 2010 Thomas Riboulet riboulet@gmail.com
-#
-#Permission is hereby granted, free of charge, to any person obtaining a copy
-#of this software and associated documentation files (the "Software"), to deal
-#in the Software without restriction, including without limitation the rights
-#to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-#copies of the Software, and to permit persons to whom the Software is
-#furnished to do so, subject to the following conditions:
-#
-#The above copyright notice and this permission notice shall be included in
-#all copies or substantial portions of the Software.
-#
-#THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-#IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-#FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-#AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-#LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-#OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-#THE SOFTWARE.
-
-
+#!/bin/env ruby
 require 'rubygems'
 require 'memcache'
 require 'optparse'
@@ -33,6 +12,7 @@ OptionParser.new do |opts|
    end
 end.parse!
 
+exit(1) unless File.exist?(config)
 # loading config
 config = YAML::load(File.open(options[:config_file]))[:config]
 
@@ -52,7 +32,7 @@ def get_files(a_dir)
     else
       if file_extensions.size > 0
         file_extensions.each do |ext|
-          files << "#{a_dir}/" + a_file if a_file ~= /#{ext}/
+          files << "#{a_dir}/" + a_file if a_file ~= /#{ext}$/
         end
       else
         files << "#{a_dir}/" + a_file
